Move experiments to its own folder
