+++ webview/src/setup/components/experiments/Experiments.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { useDispatch } from 'react-redux'
-import { showExperiments, showScmPanel } from './messages'
 import { NoData } from './NoData'
 import { NeedsGitCommit } from './NeedsGitCommit'
-import { EmptyState } from '../../shared/components/emptyState/EmptyState'
-import { IconButton } from '../../shared/components/button/IconButton'
-import { Beaker } from '../../shared/components/icons'
-import { Button } from '../../shared/components/button/Button'
-import { updateSectionCollapsed } from '../state/setupDataSlice'
+import { showExperiments, showScmPanel } from '../messages'
+import { EmptyState } from '../../../shared/components/emptyState/EmptyState'
+import { IconButton } from '../../../shared/components/button/IconButton'
+import { Beaker } from '../../../shared/components/icons'
+import { Button } from '../../../shared/components/button/Button'
+import { updateSectionCollapsed } from '../../state/setupDataSlice'
 
 type ExperimentsProps = {
   isDvcSetup: boolean
